refactor(contacts): extract MergeDialog (#166)
The field-by-field merge picker (#129) moves to
components/contacts/MergeDialog.vue with its choices, preview and
the update-then-delete API calls (stay-open-on-error is dialog
UI). Emits merged(surviving); the view still snapshots the pair,
closes the dialog before the best-effort refresh (double-merge
guard) and re-points the selection. All merge-* test-ids kept.
+4 tests.
